LayoutTests/webgl/2.0.0/conformance2/glsl3/uniform-location-length-limits-expected.txt

This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance2/glsl3/uniform-location-length-limits.html
gl.getProgramParameter(program, gl.LINK_STATUS) is true
[ FAIL: 2 ] uniformLoc should be non-null. Was null
[ FAIL: 3 ] getError expected: NO_ERROR. Was INVALID_VALUE :
getError was expected value: NO_ERROR :
gl.getProgramParameter(program, gl.LINK_STATUS) is true
getError was expected value: INVALID_VALUE :
uniformLoc is null
successfullyParsed is true
FAIL
